requirements.txt
beautifulsoup4==4.12.2
PyMuPDF==1.23.5
nltk==3.8.1
Pillow==10.0.1
pytesseract==0.3.10
tika==2.6.0
Flask
pandas
tika
python-docx
tabulate
openpyxl
python-pptx
Pillow
pytesseract
inscriptis
beautifulsoup4
nltk
torch
scikit-learn
transformers
docx2pdf